# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Test ExpressJS API",
        "version": "version 1.0"
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/test": {
            "get": {
                "summary": "GET /api/test",
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Generic"
                ]
            }
        },
        "/api/books": {
            "get": {
                "summary": "GET /api/books",
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Generic"
                ]
            }
        },
        "/api/book": {
            "put": {
                "summary": "PUT /api/:book",
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Generic"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "title",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "author",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "year",
                        "required": true,
                        "type": "number"
                    }
                ]
            },
            "get": {
                "summary": "GET /api/:book",
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Generic"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ]
            },
            "delete": {
                "summary": "DELETE /api/:book",
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Generic"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        }
    }
}